Refactor - move getFilterDefinitions() to logfilter.go
